requirements.txt
asgiref==3.9.1
Django==5.2.5
django-filter==25.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
sqlparse==0.5.3
gunicorn==20.1.0
Pillow==9.4.0
django-cors-headers==3.14.0
psycopg2-binary==2.9.5
celery==5.5.0
redis==4.6.0
drf-spectacular==0.28.0
whitenoise==6.5.0